bug 0002467 fixed: Import from autoexec_.ini to db doesn't works for key values in CP1251 and UTF-8
git-svn-id: http://miranda.googlecode.com/svn/trunk@3694 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/src/modules/database/dbini.c
@@ -336,7 +336,7 @@ static void DoAutoExec(void)
 		while(!feof(fp)) {
 			if(fgets(szLine,sizeof(szLine),fp)==NULL) break;
 			lineLength=lstrlenA(szLine);
-			while(lineLength && szLine[lineLength-1]<=' ') szLine[--lineLength]='\0';
+			while(lineLength && (BYTE)(szLine[lineLength-1])<=' ') szLine[--lineLength]='\0';
 			if(szLine[0]==';' || szLine[0]<=' ') continue;
 			if(szLine[0]=='[') {
 				char *szEnd=strchr(szLine+1,']');
